requirements.txt
# Core libraries
pandas>=1.3.0,<2.0.0
numpy>=1.20.0,<2.0.0
pyarrow>=7.0.0 # Often needed by pandas for certain formats or BQ interaction
db-dtypes>=1.0.0 # Required by pandas read_gbq for certain BigQuery types

# Machine Learning
xgboost>=1.5.0,<2.0.0
scikit-learn>=1.0.0,<1.2.0 # For splitting, metrics

# Google Cloud
google-cloud-aiplatform>=1.18.1,<1.30.0 # For Vertex AI HPT Job and Pipelines
google-cloud-bigquery>=2.34.0,<3.5.0 # For BigQuery interaction
google-cloud-storage>=2.0.0,<3.0.0 # Often needed implicitly by AI Platform
google-cloud-bigquery-storage>=2.10.0,<3.0.0 # For faster BQ downloads

# Kubeflow Pipelines (if compiling/running pipeline from same environment)
# kfp>=1.8.10,<2.0.0 # Use kfp v1 syntax based on the pipeline code
kfp==1.8.22 # Pinning a specific version known to work well

# Utilities
pyyaml>=6.0 # For reading config.yaml
hypertune>=1.0.0 # For reporting metrics to Vertex AI HPT
holidays>=0.13 # For generating holiday features

# Visualization (Optional - useful for notebooks/local runs, not strictly needed in Docker for training)
 matplotlib>=3.5.0
 seaborn>=0.11.0
 plotly>=5.5.0

# Add other specific dependencies if needed
